+++ jbpm-wb-integration/jbpm-wb-integration-client/src/main/java/org/jbpm/workbench/wi/client/editors/deployment/descriptornew/items/NamedObjectItemView.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.jbpm.workbench.wi.client.editors.deployment.descriptor2.items;
+package org.jbpm.workbench.wi.client.editors.deployment.descriptornew.items;
 
 import javax.inject.Inject;
 import javax.inject.Named;
@@ -30,8 +30,13 @@ import org.jboss.errai.ui.shared.api.annotations.EventHandler;
 import org.jboss.errai.ui.shared.api.annotations.Templated;
 
 @Templated("#root")
-public class ObjectItemView implements ObjectItemPresenter.View,
-                                       IsElement {
+public class NamedObjectItemView implements NamedObjectItemPresenter.View,
+                                            IsElement {
+
+    @Inject
+    @Named("span")
+    @DataField("name")
+    private HTMLElement name;
 
     @Inject
     @Named("span")
@@ -55,10 +60,10 @@ public class ObjectItemView implements ObjectItemPresenter.View,
     @DataField("remove-button")
     private HTMLAnchorElement removeButton;
 
-    private ObjectItemPresenter presenter;
+    private NamedObjectItemPresenter presenter;
 
     @Override
-    public void init(final ObjectItemPresenter presenter) {
+    public void init(final NamedObjectItemPresenter presenter) {
         this.presenter = presenter;
     }
 
@@ -70,6 +75,11 @@ public class ObjectItemView implements ObjectItemPresenter.View,
     @EventHandler("parameters-link")
     private void onParametersLinkClicked(final ClickEvent ignore) {
         presenter.showParametersModal();
+    }
+
+    @Override
+    public void setName(final String name) {
+        this.name.textContent = name;
     }
 
     @Override
