+++ jbpm-wb-integration/jbpm-wb-integration-client/src/main/java/org/jbpm/workbench/wi/client/editors/deployment/descriptornew/items/ObjectPresenter.java
@@ -14,13 +14,11 @@
  * limitations under the License.
  */
 
-package org.jbpm.workbench.wi.client.editors.deployment.descriptor2.model;
+package org.jbpm.workbench.wi.client.editors.deployment.descriptornew.items;
 
-import org.jboss.errai.common.client.api.annotations.Portable;
+public interface ObjectPresenter {
 
-@Portable
-public enum Resolver {
-    MVEL,
-    REFLECTION,
-    SPRING;
+    void fireChangeEvent();
+
+    void signalParameterAddedOrRemoved();
 }
