+++ jbpm-wb-integration/jbpm-wb-integration-client/src/main/java/org/jbpm/workbench/wi/client/editors/deployment/descriptornew/model/AuditMode.java
@@ -14,11 +14,13 @@
  * limitations under the License.
  */
 
-package org.jbpm.workbench.wi.client.editors.deployment.descriptor2.items;
+package org.jbpm.workbench.wi.client.editors.deployment.descriptornew.model;
 
-public interface ObjectPresenter {
+import org.jboss.errai.common.client.api.annotations.Portable;
 
-    void fireChangeEvent();
-
-    void signalParameterAddedOrRemoved();
+@Portable
+public enum AuditMode {
+    JPA,
+    JMS,
+    NONE;
 }
