+++ jbpm-wb-integration/jbpm-wb-integration-client/src/main/java/org/jbpm/workbench/wi/client/editors/deployment/descriptornew/model/PersistenceMode.java
@@ -14,14 +14,12 @@
  * limitations under the License.
  */
 
-package org.jbpm.workbench.wi.client.editors.deployment.descriptor2.model;
+package org.jbpm.workbench.wi.client.editors.deployment.descriptornew.model;
 
 import org.jboss.errai.common.client.api.annotations.Portable;
 
 @Portable
-public enum RuntimeStrategy {
-    SINGLETON,
-    PER_REQUEST,
-    PER_PROCESS_INSTANCE
+public enum PersistenceMode {
+    JPA,
+    NONE;
 }
-
