+++ jbpm-wb-integration/jbpm-wb-integration-client/src/main/java/org/jbpm/workbench/wi/client/editors/deployment/descriptornew/model/Resolver.java
@@ -14,12 +14,13 @@
  * limitations under the License.
  */
 
-package org.jbpm.workbench.wi.client.editors.deployment.descriptor2.model;
+package org.jbpm.workbench.wi.client.editors.deployment.descriptornew.model;
 
 import org.jboss.errai.common.client.api.annotations.Portable;
 
 @Portable
-public enum PersistenceMode {
-    JPA,
-    NONE;
+public enum Resolver {
+    MVEL,
+    REFLECTION,
+    SPRING;
 }
